Prevent height change in Collapse component title (#3144)
When toggling show/hide, if the title has a different background color than the content, and if there is lengthy content, you can see the title row/track grow during the animation before returning to its original size. Using `max-content` instead of `auto` in the grid-template-rows rule will lock the height of the title, so only the 2nd row (content) changes from `0fr` to `1fr`.