Color SVG fills the base content color by default
DaisyUI sets "color" to --color-base-content on all elements. In addition to all text, this colors SVGs, but only the stroke part, not the fill. The SVG fill in icons, like an icon from rails_icons (https://github.com/Rails-Designer/rails_icons) isn't colored.

This can be surprising when injecting icons anywhere in buttons or input forms, for example. I was working on a password form, and a path-only icon was colored correctly, but a fill-only icon wasn't, and I needed to set the color myself.

It might actually be a good idea to set "fill" whenever "color" gets set, too... but I figured I'd bring up the idea here.