feat: move logic for drawer required backdrop in drawer (from rootscrolllock)
This move is good from the POV of css generation because now the styles are auto-generated in case variants are used for drawer-open.

This move is not good because now we have styles that concern rootscrolllock in drawer.

The best would be if we could move these 2 blocks of code in rootscrolllock and declare them as variants as they are in drawer.