feat: scrollbar, drawer and modal improvements
* feat: add a transition for :root bg change on scrolllock to sync with backdrop

* feat: allow auto height for drawer-open

- drawer-open will be part of the page content and we do not want it to be forced to screen height (it will participate to general page height)
- if we want to force it to screen height we can use `h-screen` on `drawer-side`

* feat: reorganize layers so that :checked and drawer-open for the drawer are both in modifiers layer

This allows us to apply style when drawer visibility is checked, and revert-layer later if it is drawer-open.

Do not change scrollbar-color if .drawer-open (the scrollbar is part of the page normal flow)

* feat: move logic for drawer required backdrop in drawer (from rootscrolllock)

This move is good from the POV of css generation because now the styles are auto-generated in case variants are used for drawer-open.

This move is not good because now we have styles that concern rootscrolllock in drawer.

The best would be if we could move these 2 blocks of code in rootscrolllock and declare them as variants as they are in drawer.

* feat: solve the locality problem of the scrollock

* feat: refactor rootscrollgutter and rootscrolllock

- rootscrolllock and rootscrollgutter (if included) apply settings based on flags set by the components as needed
- modal always sets flags for overflow hidden, backdrop, ...
- drawer checked sets flags for overflow hidden, backdrop, ... in its own layer
- but drawer open reverts flags for overflow hidden, backdrop, ... in its own layer

* fix: only revert drawer setting if it was checked

* fix: add forgotten default for bg when drawer or modal are not active

* rename some variables, minor cleanups

---------

Co-authored-by: Pouya Saadeghi <pouya.saadeghi@gmail.com>